refactor: Break up liquidation receipt tests
